create separate component FormDialog, add dialogContext and provider, implement toggle FormDialog method
